requirements.txt
Unidecode==0.4.21
ckanapi==4.0
docopt==0.6.2
et-xmlfile==1.1.0
isodate==0.6.0
jsonschema==2.6.0
openpyxl==3.0.10
pydatajson==0.4.67
python-dateutil==2.8.0
requests-mock==1.9.3
rfc3987==1.3.7
unicodecsv==0.14.1
